Updated scheduling queues and scheduler API

+++ lib/task_scheduler/task_table/table.h
@@ -18,7 +18,7 @@ typedef struct entry
 {
     uint8_t id;
     void * task;
-    uint16_t * size;
+    uint8_t * size;
     struct entry * next;
 
 } task_entry_t;
